Better editing of default values.
Editing simple default values in the Datatype configuration dialog is
cumbersome: right-click and create a child using the tree and then set
properties of the new child object, including value and the link back to
the definition object. Ideally the latter should be set automatically
and remain hidden.
This change simplify the UI by hiding the AttributeValueSimple object
and by showing the default value as a direct property of the
AttributeDefinitionSimple.

Change-Id: Ife639f61f557a87fbc998ab16e3a7fe2064b1a15
Signed-off-by: François Rey <eclipse.org@francois.rey.name>