RESOLVED - bug 400951: Analyze potential side effects of setIntrinsicIDToEObjectMap https://bugs.eclipse.org/bugs/show_bug.cgi?id=400951
added adapter that creates missing IDs as soon as an EObject that has an
IDAttribute is added to the containment tree of the resource.

Lookup caches for IDs and EObjects in a resource are implemented and are
updated whenever in id is changed or EObjects with ID are removed or
added.