This fix the bug #440509.
I have updated the unit test.

Maybe we must rename the name of the class IDAdapterTests because the
class IDAdapter was deleted.

NOTE : Normally, we must not call the setUUID at all because the uuid
management is done in the resource class (so maybe a setUUID (in the
case of ReqIF setIdentifier) method must not be provided and the only
way to change an UUID must be done in a reflective way via an eSet).

Conflicts merged by michael@jastram.de:
	org.eclipse.rmf.tests.serialization/src/org/eclipse/rmf/tests/serialization/model/nodes/serialization/NodesResourceImpl.java

Change-Id: Ia981af0074c6a3d865899ddd9663aab9b012e8f7
Signed-off-by: Hussein MHANNA <hussein.mhanna@all4tec.net>