Fix the unit tests.
1. Add a boolean in the class ReqIF10ResourceImpl to give the
possibility to force the keep of the UUID.
2. In the class AbstractTestcase, I don't know why the methods are
static, because a static method can't be override, maybe a 'Utils' can
be create to put static methods.
3. In the class TC18xxHISExchangeProcessTests, add a new method
'doSaveReqIFFile' to save the file and to keep the id (the method
saveReqIFFile from the class AbstractTestCase is static so I can't
override this method).

Change-Id: I7c4f44cd0072b48c5660ed052c8308b8eb7e8960
Signed-off-by: Hussein MHANNA <hussein.mhanna@all4tec.net>